CNN/BovinaBreedPrediction/app/requirements.txt
fastapi
uvicorn
pillow
numpy
tensorflow==2.20.0  # Or your version
python-multipart
